+++ index.js
@@ -1,10 +1,10 @@
 /**
  * Internal dependencies
  */
-const RemoveSuprefluousAssetsPlugin = require("./plugins/remove-superfluous-assets");
-const webpackConfig = require("./config/webpack.config");
+const RemoveSuprefluousAssetsPlugin = require('./plugins/remove-superfluous-assets');
+const webpackConfig = require('./config/webpack.config');
 
 module.exports = {
 	RemoveSuprefluousAssetsPlugin,
-	webpackConfig
+	webpackConfig,
 };