+++ stylelint-config.js
@@ -1,0 +1,10 @@
+module.exports = {
+	extends: [
+		'stylelint-config-recommended',
+		'stylelint-config-recommended-scss',
+	],
+	plugins: ['stylelint-order'],
+	rules: {
+		'order/order': ['custom-properties', 'declarations', 'rules'],
+	},
+};